Update builder to only build CopyCapt if binary is missing.

+++ updater/builder.sh
@@ -14,13 +14,17 @@ function logOut {
 #Let's compile some required binaries.  We do this every time, just incase.
 
 #Compile CutyCapt
-logOut "Compiling CutyCapt..."
-cd ${srcroot}ext-libs/CutyCapt > /dev/null 2>&1
-qmake > /dev/null 2>&1
-make > /dev/null 2>&1
-cp CutyCapt ${scrroot} > /dev/null 2>&1
-rm CutyCapt > /dev/null 2>&1
-logOut "Skipping CutyCapt Compilation - already latest."
+if [ -f ${scrroot}CutyCapt ]
+then
+	logOut "Skipping CutyCapt Compilation - already latest."
+else
+	logOut "Compiling CutyCapt..."
+	cd ${srcroot}ext-libs/CutyCapt > /dev/null 2>&1
+	qmake > /dev/null 2>&1
+	make > /dev/null 2>&1
+	cp CutyCapt ${scrroot} > /dev/null 2>&1
+	rm CutyCapt > /dev/null 2>&1
+fi
 
 #We're updating - kill previous infodisplay
 killall -9 infodisplay > /dev/null 2>&1
@@ -41,7 +45,6 @@ cp infodisplay ${scrroot} > /dev/null 2>&1
 rm infodisplay > /dev/null 2>&1
 cp -a fonts ${scrroot}
 cp -a textures ${scrroot}
-cp -a boards ${scrroot}
 
 # Load Application
 logOut "Launching InfoDisplay..."
